refactor(upload): merge archive.go into uploadmodule.go and write unit tests
