+++ cmd/uploadmodule_test.go
@@ -2,15 +2,19 @@ package cmd
 
 import (
 	//"context"
-	//"io"
-	//"os"
-	//"path/filepath"
+	"archive/tar"
+	"compress/gzip"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
 	"testing"
 
 	//"github.com/boring-registry/boring-registry/pkg/core"
 	"github.com/boring-registry/boring-registry/pkg/module"
 
 	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/assert"
 )
 
 //type Mock struct{}
@@ -31,7 +35,7 @@ func TestModuleUploadRunner_Run(t *testing.T) {
 	validPath := t.TempDir()
 	m := &moduleUploadRunner{
 		//storage: &Mock{},
-		archive: func(_ string, _ module.Storage) error { return nil },
+		discover: func(_ string, _ module.Storage) error { return nil },
 	}
 
 	tests := []struct {
@@ -130,3 +134,151 @@ func TestModuleUploadRunner_Run(t *testing.T) {
 //
 //	return dir
 //}
+
+func TestArchiveFileHeaderName(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		name   string
+		root   string
+		path   string
+		result string
+	}{
+		{
+			name:   "top-level file in a module",
+			root:   "/tmp/boring-registry/modules/example",
+			path:   "/tmp/boring-registry/modules/example/main.tf",
+			result: "main.tf",
+		},
+		{
+			name:   "nested file in a module",
+			root:   "/tmp/boring-registry/modules/example",
+			path:   "/tmp/boring-registry/modules/example/modules/auth/main.tf",
+			result: "modules/auth/main.tf",
+		},
+		{
+			name:   "hidden file without file extension",
+			root:   "/tmp/boring-registry/modules/example",
+			path:   "/tmp/boring-registry/modules/example/.hidden",
+			result: ".hidden",
+		},
+		{
+			name:   "hidden file without recursive walk",
+			root:   ".",
+			path:   ".hidden",
+			result: ".hidden",
+		},
+		{
+			name:   "file path with parent directory",
+			root:   "../../tmp/boring-registry/modules/example",
+			path:   "../../tmp/boring-registry/modules/example/main.tf",
+			result: "main.tf",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			assert.Equal(t, tc.result, archiveFileHeaderName(tc.path, tc.root))
+		})
+	}
+
+}
+
+func TestArchiveModule(t *testing.T) {
+	t.Parallel()
+
+	type file struct {
+		content  string
+		fileMode os.FileMode
+	}
+	tests := []struct {
+		name               string
+		files              map[string]file
+		useNonExistentPath bool
+		wantErr            bool
+	}{
+		{
+			name: "archive module directory successfully",
+			files: map[string]file{
+				"main.tf":                 {content: "test content"},
+				"variables.tf":            {content: "test content"},
+				"modules/example/test.tf": {content: "nested content"},
+			},
+			wantErr: false,
+		},
+		{
+			name: "file without read permissions",
+			files: map[string]file{
+				"main.tf":                 {content: "test content"},
+				"variables.tf":            {content: "test content", fileMode: 0200}, // write-only
+				"modules/example/test.tf": {content: "nested content"},
+			},
+			wantErr: true,
+		},
+		{
+			name:               "non-existent directory",
+			useNonExistentPath: true,
+			wantErr:            true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			var dir string
+			if tt.useNonExistentPath {
+				dir = "/non/existent/path"
+			} else {
+				dir = t.TempDir()
+				// Create test files
+				for path, f := range tt.files {
+					fullPath := filepath.Join(dir, path)
+					err := os.MkdirAll(filepath.Dir(fullPath), 0755)
+					assert.NoError(t, err)
+
+					mode := os.FileMode(0644)
+					if f.fileMode != 0 {
+						mode = f.fileMode
+					}
+					err = os.WriteFile(fullPath, []byte(f.content), mode)
+					assert.NoError(t, err)
+				}
+			}
+
+			reader, err := archiveModule(dir)
+			if tt.wantErr {
+				assert.Error(t, err)
+				return
+			}
+
+			assert.NoError(t, err)
+			assert.NotNil(t, reader)
+
+			// Verify archive contents
+			gzr, err := gzip.NewReader(reader)
+			assert.NoError(t, err)
+
+			defer func() {
+				assert.NoError(t, gzr.Close())
+			}()
+
+			tr := tar.NewReader(gzr)
+			foundFiles := make(map[string]bool)
+
+			for {
+				header, err := tr.Next()
+				if err == io.EOF {
+					break
+				}
+				assert.NoError(t, err)
+				foundFiles[header.Name] = true
+			}
+
+			// Verify all test files are in archive
+			for fileName := range tt.files {
+				assert.True(t, foundFiles[fileName], fmt.Sprintf("file %s not found in archive", fileName))
+			}
+		})
+	}
+}
